fix(query-editor) Solving issue regarding duplicated sugestions in query editor.
