+++ ui/src/views/sql-query-editor/components/sql-editor-section.tsx
@@ -1,12 +1,7 @@
 
-import { useQuery } from '@apollo/client'
-import Editor, { useMonaco } from '@monaco-editor/react'
+import Editor from '@monaco-editor/react'
 import { useEffect, useRef } from 'react'
-import { GetSchemaQuery } from 'src/api-logic/graphql/generated/schema'
-import { GET_SCHEMA_INFO } from 'src/api-logic/graphql/queries/get-schema'
-import { getRecordsFromSchema } from 'src/api-logic/mappers/schema'
 import { useQueryContext, useQuerySetState } from 'src/state/contexts/query.context'
-import { sqlKeywords } from 'src/utils/monaco-editor/autocomplete/sql-keywords'
 import SchemaSection from './schema-section'
 
 type SQLEditorSectionProps = {
@@ -17,15 +12,9 @@ type SQLEditorSectionProps = {
 const SQLEditorSection: React.FC<SQLEditorSectionProps> = ({ onEnterKey }: SQLEditorSectionProps) => {
   const [{ query }] = useQueryContext()
   const { setQuery } = useQuerySetState()
-  const monaco = useMonaco()
 
   const resizeElement = useRef<HTMLDivElement | null>(null)
   const resizerElement = useRef<HTMLDivElement | null>(null)
-
-  const { data } = useQuery<GetSchemaQuery>(GET_SCHEMA_INFO, {
-    variables: { search: '' },
-    fetchPolicy: 'no-cache'
-  })
 
   /**
    * useEffect for handle resizing (horizontal divider)
@@ -75,66 +64,6 @@ const SQLEditorSection: React.FC<SQLEditorSectionProps> = ({ onEnterKey }: SQLEd
       window.removeEventListener('keydown', handleKeyDown)
     }
   }, [onEnterKey])
-
-  /**
-   * useEffect for adding autocomplete keywords
-   */
-  useEffect(() => {
-    if (monaco && data) {
-      monaco.languages.registerCompletionItemProvider('sql', {
-        provideCompletionItems: (model, position) => {
-          const word = model.getWordUntilPosition(position)
-          const range = {
-            startLineNumber: position.lineNumber,
-            endLineNumber: position.lineNumber,
-            startColumn: word.startColumn,
-            endColumn: word.endColumn,
-          }
-
-          const keywordSuggestions = sqlKeywords.map((keyword) => ({
-            label: keyword,
-            kind: monaco.languages.CompletionItemKind.Keyword,
-            documentation: keyword + ' (SQL keyword)',
-            insertText: keyword,
-            range,
-          }))
-
-          const schemaSuggestions = getRecordsFromSchema(data, 'schema').map((schema) => ({
-            label: schema,
-            kind: monaco.languages.CompletionItemKind.Function,
-            documentation: schema + ' (schema)',
-            insertText: schema,
-            range,
-          }))
-
-          const tableSuggestions = getRecordsFromSchema(data, 'tableName').map((table) => ({
-            label: table,
-            kind: monaco.languages.CompletionItemKind.Function,
-            documentation: table + ' (table)',
-            insertText: table,
-            range,
-          }))
-
-          const columnSuggestions = getRecordsFromSchema(data, 'columnName').map((column) => ({
-            label: column,
-            kind: monaco.languages.CompletionItemKind.Field,
-            documentation: column + ' (column)',
-            insertText: column,
-            range,
-          }))
-
-          return {
-            suggestions: [
-              ...keywordSuggestions,
-              ...schemaSuggestions,
-              ...tableSuggestions,
-              ...columnSuggestions,
-            ],
-          }
-        },
-      })
-    }
-  }, [monaco, data])
 
   return (
     <>
